test to access to request->payload

+++ examples/vlc-module/sdlib/vip/vip.c
@@ -272,7 +272,7 @@ vip_serialize_VM(vip_message_t *vip_pkt, uint8_t *buffer)
     return offset + index;
 }
 
-int vip_parse_message(vip_message_t *vip_pkt, uint8_t *data, uint32_t data_len)
+int vip_parse_message(vip_message_t *vip_pkt, uint8_t *data, uint16_t data_len)
 {
     //memset(vip_pkt, 0, sizeof(vip_message_t));
